# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Tobfa API",
        "contact": {
            "name": "Tobi",
            "email": "tobiaditia549@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/businessTransactions": {
            "post": {
                "description": "Create Business Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Transaction"
                ],
                "summary": "Create Business Transaction",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.BusinessTransactionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.BusinessTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/businessTransactions/{id}": {
            "get": {
                "description": "Find Business Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Transaction"
                ],
                "summary": "Find Business Transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.BusinessTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update Business Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Transaction"
                ],
                "summary": "Update Business Transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.BusinessTransactionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.BusinessTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Business Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Transaction"
                ],
                "summary": "Delete Business Transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/businessTransactions/{id}/business": {
            "get": {
                "description": "Find Business Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Transaction"
                ],
                "summary": "Find Business Transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Business ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web.BusinessTransactionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/stats/businessTransactions": {
            "get": {
                "description": "Stats Business Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "summary": "Stats Business Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dateStarted",
                        "name": "dateStarted",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dateEnded",
                        "name": "dateEnded",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "businessTransactionTypeId",
                        "name": "businessTransactionTypeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "businessTransactionItemId",
                        "name": "businessTransactionItemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "provinceId",
                        "name": "provinceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "cityId",
                        "name": "cityId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "districtId",
                        "name": "districtId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "villageId",
                        "name": "villageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "businessCategoryId",
                        "name": "businessCategoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web.BusinessTransactionStatsResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "web.BusinessCategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "web.BusinessTransactionCreateRequest": {
            "type": "object",
            "required": [
                "businessId",
                "businessTransactionItemId",
                "businessTransactionTypeId",
                "date",
                "description",
                "multiplier",
                "total"
            ],
            "properties": {
                "businessId": {
                    "type": "integer"
                },
                "businessTransactionItemId": {
                    "type": "integer"
                },
                "businessTransactionTypeId": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "multiplier": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "web.BusinessTransactionItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "web.BusinessTransactionResponse": {
            "type": "object",
            "properties": {
                "businessId": {
                    "type": "integer"
                },
                "businessTransactionItem": {
                    "$ref": "#/definitions/web.BusinessTransactionItemResponse"
                },
                "businessTransactionItemId": {
                    "type": "integer"
                },
                "businessTransactionTypeId": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "multiplier": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "web.BusinessTransactionStatsResponse": {
            "type": "object",
            "properties": {
                "averange": {
                    "type": "integer"
                },
                "businessCategory": {
                    "$ref": "#/definitions/web.BusinessCategoryResponse"
                },
                "stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.StatsResponse"
                    }
                }
            }
        },
        "web.BusinessTransactionUpdateRequest": {
            "type": "object",
            "required": [
                "businessId",
                "businessTransactionItemId",
                "businessTransactionTypeId",
                "date",
                "description",
                "multiplier",
                "total"
            ],
            "properties": {
                "businessId": {
                    "type": "integer"
                },
                "businessTransactionItemId": {
                    "type": "integer"
                },
                "businessTransactionTypeId": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "multiplier": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "web.StatsResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "web.WebResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}